started refactoring so the environment information is not hardcoded
