+++ integrationTests/src/main/java/DockerMachine.java
@@ -6,6 +6,6 @@ public class DockerMachine {
     }
 
     public String getIpAddress() {
-        return docker.readStandardOutput("docker-machine ip "+ docker.getMachineName())[0];
+        return new ProcessWrapper().readStandardOutput("docker-machine ip "+ docker.getMachineName())[0];
     }
 }
